refactor(vehicles): split application service modules
