+++ backend/internal/application/vehicle_mutation.go
@@ -1,0 +1,347 @@
+package application
+
+import (
+	"context"
+	"fmt"
+	"strings"
+	"time"
+)
+
+func (s *VehicleService) Create(ctx context.Context, input CreateVehicleInput, actorUserID string) (*Vehicle, error) {
+	input = cleanVehicleInput(input)
+	if input.Manufacturer == "" || input.Name == "" || input.Gauge == "" || input.Category == "" || input.Gattung == "" {
+		return nil, ErrVehicleValidation
+	}
+	vehicleID := randomID()
+	var err error
+	if s.imageLocalizer != nil && len(input.Images) > 0 {
+		input.Images, err = s.imageLocalizer(ctx, vehicleID, input.Images)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	tx, err := s.db.BeginTx(ctx, nil)
+	if err != nil {
+		return nil, fmt.Errorf("begin create vehicle: %w", err)
+	}
+	defer func() {
+		if err != nil {
+			_ = tx.Rollback()
+		}
+	}()
+
+	if input.InventoryNumber == "" {
+		input.InventoryNumber, err = s.nextInventoryNumber(ctx, tx, input.Category)
+		if err != nil {
+			return nil, err
+		}
+	} else if err = s.ensureInventoryNumberAvailable(ctx, tx, input.InventoryNumber, ""); err != nil {
+		return nil, err
+	}
+
+	now := time.Now().UTC().Format(time.RFC3339)
+	vehicle := Vehicle{
+		ID:                        vehicleID,
+		InventoryNumber:           input.InventoryNumber,
+		Manufacturer:              input.Manufacturer,
+		ArticleNumber:             input.ArticleNumber,
+		ArticleSourceURL:          input.ArticleSourceURL,
+		Name:                      input.Name,
+		Gauge:                     input.Gauge,
+		Epoch:                     input.Epoch,
+		RailwayCompany:            input.RailwayCompany,
+		Category:                  input.Category,
+		Gattung:                   input.Gattung,
+		Description:               input.Description,
+		Series:                    input.Series,
+		VehicleNumber:             input.VehicleNumber,
+		Digital:                   input.Digital,
+		DigitalDecoderNumber:      input.DigitalDecoderNumber,
+		DTDecoder:                 input.DTDecoder,
+		DTDecoderNumber:           input.DTDecoderNumber,
+		DecoderType:               input.DecoderType,
+		ExhibitionReady:           input.ExhibitionReady,
+		Exhibition:                input.Exhibition,
+		ABCBrakes:                 input.ABCBrakes,
+		EAN:                       input.EAN,
+		ProductionPeriod:          input.ProductionPeriod,
+		ListPrice:                 input.ListPrice,
+		AcquisitionType:           input.AcquisitionType,
+		AcquiredFrom:              input.AcquiredFrom,
+		PurchasePrice:             input.PurchasePrice,
+		PurchaseDate:              input.PurchaseDate,
+		StorageLocation:           input.StorageLocation,
+		StorageDetails:            input.StorageDetails,
+		Condition:                 input.Condition,
+		ConditionDetails:          input.ConditionDetails,
+		Packaging:                 input.Packaging,
+		LengthMM:                  input.LengthMM,
+		WeightG:                   input.WeightG,
+		Color:                     input.Color,
+		Lettering:                 input.Lettering,
+		Load:                      input.Load,
+		Interior:                  input.Interior,
+		Axles:                     input.Axles,
+		AxleCount:                 input.AxleCount,
+		TractionTireCount:         input.TractionTireCount,
+		Wheelset:                  input.Wheelset,
+		CouplingSame:              input.CouplingSame,
+		CouplingFront:             input.CouplingFront,
+		CouplingRear:              input.CouplingRear,
+		PowerPickup:               input.PowerPickup,
+		Adapter:                   input.Adapter,
+		DriveEnabled:              input.DriveEnabled,
+		DriveDescription:          input.DriveDescription,
+		HeadlightsEnabled:         input.HeadlightsEnabled,
+		HeadlightsDescription:     input.HeadlightsDescription,
+		LightingEnabled:           input.LightingEnabled,
+		LightingDescription:       input.LightingDescription,
+		SoundGeneratorEnabled:     input.SoundGeneratorEnabled,
+		SoundGeneratorDescription: input.SoundGeneratorDescription,
+		SmokeGeneratorEnabled:     input.SmokeGeneratorEnabled,
+		SmokeGeneratorDescription: input.SmokeGeneratorDescription,
+		AdditionalInfo:            input.AdditionalInfo,
+		QRCodeEnabled:             input.QRCodeEnabled,
+		Images:                    vehicleImagesFromInput(vehicleID, input.Images, now),
+		CreatedAt:                 now,
+		UpdatedAt:                 now,
+	}
+
+	if _, err = tx.ExecContext(ctx, `
+INSERT INTO vehicles(
+  id, inventory_number, manufacturer, article_number, article_source_url, name, gauge, epoch, railway_company, category, gattung,
+  description, series, vehicle_number, digital, digital_decoder_number, dt_decoder, dt_decoder_number, decoder_type,
+  exhibition_ready, exhibition, abc_brakes, ean, production_period, list_price,
+  acquisition_type, acquired_from, purchase_price, purchase_date, storage_location, storage_details, condition, condition_details, packaging,
+  length_mm, weight_g, color, lettering, load, interior, axles, axle_count, traction_tire_count, wheelset,
+  coupling_same, coupling_front, coupling_rear, power_pickup, adapter,
+  drive_enabled, drive_description, headlights_enabled, headlights_description, lighting_enabled, lighting_description,
+  sound_generator_enabled, sound_generator_description, smoke_generator_enabled, smoke_generator_description,
+  additional_info, qr_code_enabled, created_at, updated_at
+)
+VALUES(?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
+`, vehicle.ID, vehicle.InventoryNumber, vehicle.Manufacturer, vehicle.ArticleNumber, vehicle.ArticleSourceURL, vehicle.Name, vehicle.Gauge, vehicle.Epoch, vehicle.RailwayCompany, vehicle.Category, vehicle.Gattung, vehicle.Description, vehicle.Series, vehicle.VehicleNumber, boolToInt(vehicle.Digital), vehicle.DigitalDecoderNumber, boolToInt(vehicle.DTDecoder), vehicle.DTDecoderNumber, vehicle.DecoderType, boolToInt(vehicle.ExhibitionReady), boolToInt(vehicle.Exhibition), boolToInt(vehicle.ABCBrakes), vehicle.EAN, vehicle.ProductionPeriod, vehicle.ListPrice, vehicle.AcquisitionType, vehicle.AcquiredFrom, vehicle.PurchasePrice, vehicle.PurchaseDate, vehicle.StorageLocation, vehicle.StorageDetails, vehicle.Condition, vehicle.ConditionDetails, vehicle.Packaging, vehicle.LengthMM, vehicle.WeightG, vehicle.Color, vehicle.Lettering, vehicle.Load, vehicle.Interior, vehicle.Axles, vehicle.AxleCount, vehicle.TractionTireCount, vehicle.Wheelset, boolToInt(vehicle.CouplingSame), vehicle.CouplingFront, vehicle.CouplingRear, vehicle.PowerPickup, vehicle.Adapter, boolToInt(vehicle.DriveEnabled), vehicle.DriveDescription, boolToInt(vehicle.HeadlightsEnabled), vehicle.HeadlightsDescription, boolToInt(vehicle.LightingEnabled), vehicle.LightingDescription, boolToInt(vehicle.SoundGeneratorEnabled), vehicle.SoundGeneratorDescription, boolToInt(vehicle.SmokeGeneratorEnabled), vehicle.SmokeGeneratorDescription, vehicle.AdditionalInfo, boolToInt(vehicle.QRCodeEnabled), vehicle.CreatedAt, vehicle.UpdatedAt); err != nil {
+		return nil, fmt.Errorf("insert vehicle: %w", err)
+	}
+
+	if _, err = tx.ExecContext(ctx, `
+INSERT INTO audit_logs(id, actor_user_id, action, target_type, target_id, created_at, details_json)
+VALUES(?, ?, 'VehicleCreated', 'vehicle', ?, ?, '{}')
+`, randomID(), actorUserID, vehicle.ID, now); err != nil {
+		return nil, fmt.Errorf("write vehicle audit log: %w", err)
+	}
+	if err = saveVehicleImages(ctx, tx, vehicle.ID, input.Images, now); err != nil {
+		return nil, err
+	}
+
+	if err = tx.Commit(); err != nil {
+		return nil, fmt.Errorf("commit create vehicle: %w", err)
+	}
+	images, err := s.loadVehicleImages(ctx, vehicle.ID)
+	if err != nil {
+		return nil, err
+	}
+	vehicle.Images = images
+
+	return &vehicle, nil
+}
+
+func (s *VehicleService) Update(ctx context.Context, id string, input CreateVehicleInput, actorUserID string) (*Vehicle, error) {
+	id = strings.TrimSpace(id)
+	if id == "" {
+		return nil, ErrVehicleNotFound
+	}
+
+	existing, err := s.get(ctx, id)
+	if err != nil {
+		return nil, err
+	}
+
+	replaceImages := input.Images != nil
+	input = cleanVehicleInput(input)
+	if input.InventoryNumber == "" {
+		input.InventoryNumber = existing.InventoryNumber
+	}
+	if input.Manufacturer == "" || input.Name == "" || input.Gauge == "" || input.Category == "" || input.Gattung == "" {
+		return nil, ErrVehicleValidation
+	}
+	if s.imageLocalizer != nil && replaceImages && len(input.Images) > 0 {
+		input.Images, err = s.imageLocalizer(ctx, id, input.Images)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	now := time.Now().UTC().Format(time.RFC3339)
+	vehicle := Vehicle{
+		ID:                        id,
+		InventoryNumber:           input.InventoryNumber,
+		Manufacturer:              input.Manufacturer,
+		ArticleNumber:             input.ArticleNumber,
+		ArticleSourceURL:          input.ArticleSourceURL,
+		Name:                      input.Name,
+		Gauge:                     input.Gauge,
+		Epoch:                     input.Epoch,
+		RailwayCompany:            input.RailwayCompany,
+		Category:                  input.Category,
+		Gattung:                   input.Gattung,
+		Description:               input.Description,
+		Series:                    input.Series,
+		VehicleNumber:             input.VehicleNumber,
+		Digital:                   input.Digital,
+		DigitalDecoderNumber:      input.DigitalDecoderNumber,
+		DTDecoder:                 input.DTDecoder,
+		DTDecoderNumber:           input.DTDecoderNumber,
+		DecoderType:               input.DecoderType,
+		ExhibitionReady:           input.ExhibitionReady,
+		Exhibition:                input.Exhibition,
+		ABCBrakes:                 input.ABCBrakes,
+		EAN:                       input.EAN,
+		ProductionPeriod:          input.ProductionPeriod,
+		ListPrice:                 input.ListPrice,
+		AcquisitionType:           input.AcquisitionType,
+		AcquiredFrom:              input.AcquiredFrom,
+		PurchasePrice:             input.PurchasePrice,
+		PurchaseDate:              input.PurchaseDate,
+		StorageLocation:           input.StorageLocation,
+		StorageDetails:            input.StorageDetails,
+		Condition:                 input.Condition,
+		ConditionDetails:          input.ConditionDetails,
+		Packaging:                 input.Packaging,
+		LengthMM:                  input.LengthMM,
+		WeightG:                   input.WeightG,
+		Color:                     input.Color,
+		Lettering:                 input.Lettering,
+		Load:                      input.Load,
+		Interior:                  input.Interior,
+		Axles:                     input.Axles,
+		AxleCount:                 input.AxleCount,
+		TractionTireCount:         input.TractionTireCount,
+		Wheelset:                  input.Wheelset,
+		CouplingSame:              input.CouplingSame,
+		CouplingFront:             input.CouplingFront,
+		CouplingRear:              input.CouplingRear,
+		PowerPickup:               input.PowerPickup,
+		Adapter:                   input.Adapter,
+		DriveEnabled:              input.DriveEnabled,
+		DriveDescription:          input.DriveDescription,
+		HeadlightsEnabled:         input.HeadlightsEnabled,
+		HeadlightsDescription:     input.HeadlightsDescription,
+		LightingEnabled:           input.LightingEnabled,
+		LightingDescription:       input.LightingDescription,
+		SoundGeneratorEnabled:     input.SoundGeneratorEnabled,
+		SoundGeneratorDescription: input.SoundGeneratorDescription,
+		SmokeGeneratorEnabled:     input.SmokeGeneratorEnabled,
+		SmokeGeneratorDescription: input.SmokeGeneratorDescription,
+		AdditionalInfo:            input.AdditionalInfo,
+		QRCodeEnabled:             input.QRCodeEnabled,
+		Images:                    vehicleImagesFromInput(id, input.Images, now),
+		CreatedAt:                 existing.CreatedAt,
+		UpdatedAt:                 now,
+	}
+
+	tx, err := s.db.BeginTx(ctx, nil)
+	if err != nil {
+		return nil, fmt.Errorf("begin update vehicle: %w", err)
+	}
+	defer func() {
+		if err != nil {
+			_ = tx.Rollback()
+		}
+	}()
+
+	if vehicle.InventoryNumber != existing.InventoryNumber {
+		if err = s.ensureInventoryNumberAvailable(ctx, tx, vehicle.InventoryNumber, vehicle.ID); err != nil {
+			return nil, err
+		}
+	}
+
+	result, err := tx.ExecContext(ctx, `
+UPDATE vehicles
+SET inventory_number=?, manufacturer=?, article_number=?, article_source_url=?, name=?, gauge=?, epoch=?, railway_company=?, category=?, gattung=?,
+    description=?, series=?, vehicle_number=?, digital=?, digital_decoder_number=?, dt_decoder=?, dt_decoder_number=?, decoder_type=?,
+    exhibition_ready=?, exhibition=?, abc_brakes=?, ean=?, production_period=?, list_price=?,
+    acquisition_type=?, acquired_from=?, purchase_price=?, purchase_date=?, storage_location=?, storage_details=?, condition=?, condition_details=?, packaging=?,
+    length_mm=?, weight_g=?, color=?, lettering=?, load=?, interior=?, axles=?, axle_count=?, traction_tire_count=?, wheelset=?,
+    coupling_same=?, coupling_front=?, coupling_rear=?, power_pickup=?, adapter=?,
+    drive_enabled=?, drive_description=?, headlights_enabled=?, headlights_description=?, lighting_enabled=?, lighting_description=?,
+    sound_generator_enabled=?, sound_generator_description=?, smoke_generator_enabled=?, smoke_generator_description=?,
+    additional_info=?, qr_code_enabled=?, updated_at=?
+WHERE id=?
+`, vehicle.InventoryNumber, vehicle.Manufacturer, vehicle.ArticleNumber, vehicle.ArticleSourceURL, vehicle.Name, vehicle.Gauge, vehicle.Epoch, vehicle.RailwayCompany, vehicle.Category, vehicle.Gattung, vehicle.Description, vehicle.Series, vehicle.VehicleNumber, boolToInt(vehicle.Digital), vehicle.DigitalDecoderNumber, boolToInt(vehicle.DTDecoder), vehicle.DTDecoderNumber, vehicle.DecoderType, boolToInt(vehicle.ExhibitionReady), boolToInt(vehicle.Exhibition), boolToInt(vehicle.ABCBrakes), vehicle.EAN, vehicle.ProductionPeriod, vehicle.ListPrice, vehicle.AcquisitionType, vehicle.AcquiredFrom, vehicle.PurchasePrice, vehicle.PurchaseDate, vehicle.StorageLocation, vehicle.StorageDetails, vehicle.Condition, vehicle.ConditionDetails, vehicle.Packaging, vehicle.LengthMM, vehicle.WeightG, vehicle.Color, vehicle.Lettering, vehicle.Load, vehicle.Interior, vehicle.Axles, vehicle.AxleCount, vehicle.TractionTireCount, vehicle.Wheelset, boolToInt(vehicle.CouplingSame), vehicle.CouplingFront, vehicle.CouplingRear, vehicle.PowerPickup, vehicle.Adapter, boolToInt(vehicle.DriveEnabled), vehicle.DriveDescription, boolToInt(vehicle.HeadlightsEnabled), vehicle.HeadlightsDescription, boolToInt(vehicle.LightingEnabled), vehicle.LightingDescription, boolToInt(vehicle.SoundGeneratorEnabled), vehicle.SoundGeneratorDescription, boolToInt(vehicle.SmokeGeneratorEnabled), vehicle.SmokeGeneratorDescription, vehicle.AdditionalInfo, boolToInt(vehicle.QRCodeEnabled), vehicle.UpdatedAt, vehicle.ID)
+	if err != nil {
+		return nil, fmt.Errorf("update vehicle: %w", err)
+	}
+
+	affected, err := result.RowsAffected()
+	if err != nil {
+		return nil, fmt.Errorf("read update result: %w", err)
+	}
+	if affected == 0 {
+		_ = tx.Rollback()
+		return nil, ErrVehicleNotFound
+	}
+
+	if vehicle.InventoryNumber != existing.InventoryNumber {
+		if _, err = tx.ExecContext(ctx, `
+INSERT INTO inventory_number_history(id, vehicle_id, old_number, new_number, changed_by_user_id, changed_at, reason)
+VALUES(?, ?, ?, ?, ?, ?, 'manual_update')
+`, randomID(), vehicle.ID, existing.InventoryNumber, vehicle.InventoryNumber, actorUserID, now); err != nil {
+			return nil, fmt.Errorf("write inventory number history: %w", err)
+		}
+	}
+
+	if _, err = tx.ExecContext(ctx, `
+INSERT INTO audit_logs(id, actor_user_id, action, target_type, target_id, created_at, details_json)
+VALUES(?, ?, 'VehicleUpdated', 'vehicle', ?, ?, '{}')
+`, randomID(), actorUserID, vehicle.ID, now); err != nil {
+		return nil, fmt.Errorf("write vehicle audit log: %w", err)
+	}
+	if replaceImages {
+		if err = saveVehicleImages(ctx, tx, vehicle.ID, input.Images, now); err != nil {
+			return nil, err
+		}
+	}
+
+	if err = tx.Commit(); err != nil {
+		return nil, fmt.Errorf("commit update vehicle: %w", err)
+	}
+	images, err := s.loadVehicleImages(ctx, vehicle.ID)
+	if err != nil {
+		return nil, err
+	}
+	vehicle.Images = images
+	attachments, err := s.loadVehicleAttachments(ctx, vehicle.ID)
+	if err != nil {
+		return nil, err
+	}
+	vehicle.Attachments = attachments
+	maintenance, err := s.loadVehicleMaintenance(ctx, vehicle.ID)
+	if err != nil {
+		return nil, err
+	}
+	vehicle.Maintenance = maintenance
+	functions, err := s.loadVehicleFunctions(ctx, vehicle.ID)
+	if err != nil {
+		return nil, err
+	}
+	vehicle.Functions = functions
+	cvValues, err := s.loadVehicleCVValues(ctx, vehicle.ID)
+	if err != nil {
+		return nil, err
+	}
+	vehicle.CVValues = cvValues
+	cvFiles, err := s.loadVehicleCVFiles(ctx, vehicle.ID)
+	if err != nil {
+		return nil, err
+	}
+	vehicle.CVFiles = cvFiles
+	externalMappings, err := s.loadVehicleExternalMappings(ctx, id)
+	if err != nil {
+		return nil, err
+	}
+	vehicle.ExternalMappings = externalMappings
+
+	return &vehicle, nil
+}
